requirements.txt
pre-commit==4.0.1 ; python_version >= "3.13" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.13" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.13" and python_version < "4.0"
